#254 Extracting and testing querystring intepretation methods
